+++ trible/src/cli/pile/collection/maintenance_counts.rs
@@ -269,6 +269,10 @@ impl<R: CollectionRead> CollectionRead for Counted<R> {
         self.inner.records()
     }
 
+    fn collections(&self) -> Result<Vec<CollectionHandle>, Self::RecordsError> {
+        self.inner.collections()
+    }
+
     fn select_records(
         &self,
         selectors: &BTreeSet<CollectionRecordSelector>,