+++ triblespace-net/src/host/observation_tests.rs
@@ -170,6 +170,10 @@ impl<R: CollectionRead> CollectionRead for CountedSnapshot<R> {
         self.inner.records()
     }
 
+    fn collections(&self) -> Result<Vec<CollectionHandle>, Self::RecordsError> {
+        self.inner.collections()
+    }
+
     fn select_records(
         &self,
         selectors: &BTreeSet<CollectionRecordSelector>,